+++ Changelog.rst
@@ -10,7 +10,9 @@ New features
 * For CBC/CFB/OFB/CTR, `encrypt` and `decrypt` cannot be intermixed.
 * For CBC/CFB/OFB, the cipher object has both `IV` and `iv` attributes.
   `new` accepts `IV` as well as `iv` as parameter.
-* Dedicated tests for CBC/CFB/OFB/CTR, including NIST test vectors
+* For CFB/OPENPGP, `encrypt` and `decrypt` do not require the plaintext
+  or ciphertext pieces to have length multiple of the CFB segment size.
+* Dedicated tests for CBC/CFB/OFB/CTR/CCM/OPENPGP, including NIST test vectors
 * CCM cipher object has attribute `nonce`.
 * CCM cipher checks if the declared lengths of the associated data and of the
   message match the actual ones.